test: import viem v2 directly
