--- environments/v2/index.ts
@@ -1,10 +1,0 @@
-export {
-  type Client as ClientV2,
-  type PublicClient as PublicClientV2,
-  type Transport as TransportV2,
-  createClient as createClientV2,
-  http as httpV2,
-  publicActions as publicActionsV2,
-} from 'viem-v2'
-export { privateKeyToAccount as privateKeyToAccountV2 } from 'viem-v2/accounts'
-export { mainnet as mainnetV2 } from 'viem-v2/chains'
