+++ test/tsconfig.json
@@ -2,7 +2,6 @@
   "extends": "../tsconfig.base.json",
   "include": [
     ".",
-    "../environments/v2/index.ts",
     "../src",
     "../contracts/generated.ts"
   ],
